feat(layout): migrate floating controls dock to global RootLayout, making it visible on all static routes
