+++ src/app/layout.tsx
@@ -8,6 +8,7 @@ import { ThemeProvider } from "@/lib/theme-context";
 import { ToastProvider } from "@/components/Toast";
 import Script from "next/script";
 import { HomeStructuredData } from "@/components/StructuredData";
+import GlobalFloatingDock from "@/components/GlobalFloatingDock";
 
 const geistSans = Geist({
   variable: "--font-geist-sans",
@@ -145,6 +146,7 @@ export default function RootLayout({
               <ToastProvider>
                 <Header />
                 {children}
+                <GlobalFloatingDock />
               </ToastProvider>
             </SearchProvider>
           </TabProvider>
